updated WorkerThread class in qthreads.py
added documentation for valid parameters

added validation of parameters

changed terminate() calls to use quit() and exit()  as terminate() would cause the program to hang upon exiting. Presumably the worker threads were not exiting properly so the main gui thread could not quit.

added rudimentary task for loading LEED data. This needs to be changed in the future to support loading multiple data types possibly by adding a new parameter to be passed in from the main gui thread.