fixed loading of TIFF/PNG
added ability on CONFIG tab to swap the byte order of a dataset from little endian to big endian or vice versa

This is useful if the data's endianness is unknown. The program will default to reading .dat files in little endian format but the data can then be swapped if it does not look correct.

The beam shifting algorithm seems to not be working correctly for the newest dataset - unknown at this time why that is the case