fixed matplotlib errors from upgrade to pyqt5
now program gets all the way up to loading the message Console but crashes due to unhandled python exception before displaying the main window

When pyqt5 encounters an unhandled exception it calls QFatal which sends the abort signal. The code exits with code 134 (SIGABORT)

This means that the normal python stak trace does not execute for said exception.

At this point, the debugging process will take too long.

Thus currently the upgrade to pyqt5 is on hold